+++ examples/TestApi/Models/Book.cs
@@ -1,0 +1,11 @@
+using KurzSharp;
+
+namespace TestApi.Models;
+
+[RestApi]
+public partial class Book
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public string Author { get; set; }
+}