+++ examples/TestApi/Models/User.cs
@@ -1,0 +1,11 @@
+using KurzSharp;
+
+namespace TestApi.Models;
+
+[GrpcApi]
+public partial class User
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public string Address { get; set; }
+}